Renamed _thread.py
