+++ pyghthouse/ph.py
@@ -1,7 +1,7 @@
 from signal import signal, SIGINT
 
 from .data.canvas import PyghthouseCanvas
-from .thread import PHThread
+from ._thread import PHThread
 from .connection.wsconnector import VerbosityLevel
 
 # TODO: Adjust description and example for more clarity
